+++ android/src/main/kotlin/com/example/edge_gen_ai/EdgeGenAIPlugin.kt
@@ -1,4 +1,4 @@
-package com.example.edge_ai
+package com.example.edge_gen_ai
 
 import android.content.Context
 import android.graphics.BitmapFactory