+++ test/edge_gen_ai_method_channel_test.dart
@@ -1,7 +1,7 @@
 import 'dart:typed_data';
 
-import 'package:edge_ai/edge_gen_ai_method_channel.dart';
-import 'package:edge_ai/src/messages.g.dart';
+import 'package:edge_gen_ai/edge_gen_ai_method_channel.dart';
+import 'package:edge_gen_ai/src/messages.g.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 
@@ -14,7 +14,7 @@ void main() {
 
   void mockChannel(String method, Object? response) {
     messenger.setMockMessageHandler(
-      'dev.flutter.pigeon.edge_ai.EdgeGenAIHostApi.$method',
+      'dev.flutter.pigeon.edge_gen_ai.EdgeGenAIHostApi.$method',
       (ByteData? message) async {
         return EdgeGenAIHostApi.pigeonChannelCodec.encodeMessage(<Object?>[
           response,
@@ -25,7 +25,7 @@ void main() {
 
   void unmockChannel(String method) {
     messenger.setMockMessageHandler(
-      'dev.flutter.pigeon.edge_ai.EdgeGenAIHostApi.$method',
+      'dev.flutter.pigeon.edge_gen_ai.EdgeGenAIHostApi.$method',
       null,
     );
   }
@@ -43,7 +43,7 @@ void main() {
 
   test('downloadModel listens to the feature-specific channel', () async {
     const channelName =
-        'dev.flutter.pigeon.edge_ai.EdgeGenAIEventApi.summarizationDownloadProgress';
+        'dev.flutter.pigeon.edge_gen_ai.EdgeGenAIEventApi.summarizationDownloadProgress';
     final progress = EdgeGenAIDownloadProgress(
       status: EdgeGenAIDownloadStatus.completed,
     );
@@ -67,9 +67,9 @@ void main() {
 
   test('generateContent', () async {
     const hostChannel =
-        'dev.flutter.pigeon.edge_ai.EdgeGenAIHostApi.startGenerateContent';
+        'dev.flutter.pigeon.edge_gen_ai.EdgeGenAIHostApi.startGenerateContent';
     const eventChannel =
-        'dev.flutter.pigeon.edge_ai.EdgeGenAIEventApi.generateContentChunk';
+        'dev.flutter.pigeon.edge_gen_ai.EdgeGenAIEventApi.generateContentChunk';
 
     messenger.setMockMessageHandler(hostChannel, (ByteData? message) async {
       return EdgeGenAIHostApi.pigeonChannelCodec.encodeMessage(<Object?>[null]);